==> bias_correction_procedure/Step1_bias_correction_value.ipynb <==
from netCDF4 import Dataset
import pandas as pd
import numpy as np
import os
import shutil
import re
import netCDF4 as nc
import xarray as xr
from pandas import Series, DataFrame

# CMIP6 from 2011 to 2020

ds_cmip6 = xr.open_dataset('t2_1120_ssp245_lu.nc')
t2_1120_ssp245_lu = ds_cmip6['t2_1120_ssp245_lu']

#multiindex
month_hour_idx = pd.MultiIndex.from_arrays([t2_1120_ssp245_lu['Time.month'], t2_1120_ssp245_lu['Time.hour']])
t2_1120_ssp245_lu.coords['month_hour'] = ('Time', month_hour_idx)
t2_1120_ssp245_lu_month_hour = t2_1120_ssp245_lu.groupby('month_hour').mean()
#unstack
t2_1120_ssp245_lu_month_hour = t2_1120_ssp245_lu_month_hour.unstack("month_hour")
#rename dimension
t2_1120_ssp245_lu_month_hour = t2_1120_ssp245_lu_month_hour.rename({'month_hour_level_0': 'month','month_hour_level_1': 'hour'})
#reorder dimension
t2_1120_ssp245_lu_month_hour = t2_1120_ssp245_lu_month_hour.transpose("month", "hour", "south_north","west_east")

ds_1 = xr.Dataset({'t2_1120_ssp245_lu_month_hour': t2_1120_ssp245_lu_month_hour})
ds_1.to_netcdf('t2_1120_ssp245_lu_month_hour.nc')

# ERA5_1120

ds_era5 = xr.open_dataset('t2_1120_era5.nc')
t2_1120_era5 = ds_era5['t2_1120_era5']

#multiindex
month_hour_idx_2 = pd.MultiIndex.from_arrays([t2_1120_era5['Time.month'], t2_1120_era5['Time.hour']])
t2_1120_era5.coords['month_hour'] = ('Time', month_hour_idx_2)
t2_1120_era5_month_hour = t2_1120_era5.groupby('month_hour').mean()
#unstack
t2_1120_era5_month_hour = t2_1120_era5_month_hour.unstack("month_hour")
#rename dimension
t2_1120_era5_month_hour = t2_1120_era5_month_hour.rename({'month_hour_level_0': 'month','month_hour_level_1': 'hour'})
#reorder dimension
t2_1120_era5_month_hour = t2_1120_era5_month_hour.transpose("month", "hour", "south_north","west_east")

ds_2 = xr.Dataset({'t2_1120_era5_month_hour': t2_1120_era5_month_hour})
ds_2.to_netcdf('t2_1120_era5_month_hour.nc')

# Diffrence between two datasets

era5-cmip6

bias_era_cmip = t2_1120_era5_month_hour - t2_1120_ssp245_lu_month_hour

ds_3 = xr.Dataset({'bias_era5_cmip6_1120': bias_era_cmip})
ds_3.to_netcdf('bias_era5_cmip6_1120.nc')
